+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+- fix: prevent duplicate application lifecycle events ([#354](https://github.com/PostHog/posthog-ios/pull/354))
+
 ## 3.26.1 - 2025-05-30
 
 - fix: clear cached flags if remote config response hasFeatureFlags is false ([#347](https://github.com/PostHog/posthog-ios/pull/347))